+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Ability to import files by sharing them to the gallery
 - Ability to connect to the demo library by long clicking the connection screen title
+- Silent correction of PhotoPrism library page URLs during the connection.
+  For example, if pasted `http://photoprism.local/library/login` as a library root URL, 
+  it is treated as `http://photoprism.local`
 
 ### Fixed
 - "Done" button on the search people selection screen covering the last person